backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
pymupdf
pinecone-client
langchain-pinecone
langchain-huggingface
langchain-community
langchain-text-splitters
sentence-transformers
rank_bm25
upstash-redis
python-dotenv
flashrank
websockets


